// src/shared/api/data-contracts.ts
export interface HttpErrorBody {
  /** Error message */
  reason: string;
}
// ...
export interface Sticker {
  /**
   * Sticker id (send to chat with WS)
   * @format integer
   * @example 123
   */
  id?: number;
  /**
   * Url for sticker resource(image)
   * @example "/stickers/2346-dfsg-425-sdfs/14534.gif"
   */
  path?: string;
}

export interface StickerPack {
  /**
   * Sticker pack title
   * @example "pack-title"
   */
  title?: string;
  /**
   * User id that created this pack
   * @format integer
   * @example 123
   */
  user_id?: number;
  stickers?: string[];
}

export interface StickerPacksResponse {
  /** StickerPacks */
  data?: StickerPack[];
}

export interface StickersResponse {
  /** Stickers */
  data?: Sticker[];
}